Implemented mutex sharing
This is the replacement for the custom callback.
Due to spi_master thread-safety issues, this is a
workaround to allow sharing a bus-wide
mutex amongst all devices. Due to crashes related
to adding a new device when another device is
already doing polling transactions on it, the mutex
will be used for all kinds of operations on the bus
including adding/removing devices, as well as the
transactions themselves.