Fixed potential watchdog triggering
The `while` would cause the task to spin on the
core it's running on for as long as there's data
available in W5100's recv ring buffer.
Theoretically, this change can make data
reception slower. However, it's highly unlikely
this will be a relevant bottleneck, as it'll be
dwarfed by SPI speeds and firmware overhead. If
it ever comes down to be the case that this is a
relevant bottleneck, one's better off increasing
the task's priority so the chances of it getting
selected to run consecutively increases.