Lots of changes
Reorganized the component around the idea of
improving overall encapsulation, publicly exposing
the minimum necessary for interacting with it and
making internal APIs private.

Also improved separation of concerns within the
component itself, like removing direct
dependencies on lower level details of W5100 by
the esp32_mac implementation.

Added a more straightforward configuration
structure for W5100 that allows passing details
to the lower levels of the driver for
configuration. It will be expanded in the future
to expose more control over the driver.

Also improved the API contracts with the usage of
`const` wherever it fits within the W5100 code
(the lower half of the driver).

The lower level component of the driver will now
be allocated dinamically to allow for recycling
of the memory when the driver is deinitialized
(planned feature).