Fix + Feat
Updated the code that sets the ethernet hostname
on adapter initialization. This fixes a possible
issue with setting the hostname through code not
overriding the one set through Kconfig.

Added a conditional stack size increase for the
`emac_w5100_task` of 1 KiB that depends on
whether dumping data received from W5100 is
enabled or not.

Fixed a potential error with the receiving
function when RX dumping is enabled.